requirements.txt
SpeechRecognition==3.8.1
PyAudio==0.2.11
gtts==2.2.3
playsound==1.3.0
openai==0.27.0
googlesearch-python==1.0.0
beautifulsoup4==4.11.1
requests==2.28.1
